+++ src/actions/actions.js
@@ -258,7 +258,7 @@ export function evaluateNotebook(evalQueueInstance = evalQueue) {
   };
 }
 
-function getNotebookSaveRequestOptions(state, options = undefined) {
+export function getNotebookSaveRequestOptions(state, options = undefined) {
   const data = {
     title: state.title,
     content: state.jsmd
@@ -273,9 +273,14 @@ function getNotebookSaveRequestOptions(state, options = undefined) {
   return postRequestOptions;
 }
 
-function saveNotebookRequest(url, postRequestOptions, dispatch) {
-  return fetchWithCSRFTokenAndJSONContent(url, postRequestOptions).then(
-    response => {
+export function saveNotebookRequest(
+  url,
+  postRequestOptions,
+  dispatch,
+  appMsg = true
+) {
+  return fetchWithCSRFTokenAndJSONContent(url, postRequestOptions)
+    .then(response => {
       if (!response.ok) {
         return response
           .json()
@@ -303,8 +308,18 @@ function saveNotebookRequest(url, postRequestOptions, dispatch) {
           });
       }
       return response.json();
-    }
-  );
+    })
+    .catch(err => {
+      if (appMsg) {
+        dispatch(
+          updateAppMessages({
+            message: `Error Saving Notebook.`,
+            messageType: "ERROR_SAVING_NOTEBOOK"
+          })
+        );
+      }
+      throw new Error(err.message);
+    });
 }
 
 export function createNewNotebookOnServer(options = { forkedFrom: undefined }) {
@@ -313,7 +328,11 @@ export function createNewNotebookOnServer(options = { forkedFrom: undefined }) {
     const postRequestOptions = getNotebookSaveRequestOptions(state, {
       forkedFrom: options.forkedFrom
     });
-    saveNotebookRequest("/api/v1/notebooks/", postRequestOptions, dispatch)
+    return saveNotebookRequest(
+      "/api/v1/notebooks/",
+      postRequestOptions,
+      dispatch
+    )
       .then(json => {
         const message = "Notebook saved to server";
         dispatch(
@@ -332,35 +351,37 @@ export function createNewNotebookOnServer(options = { forkedFrom: undefined }) {
   };
 }
 
-export function saveNotebookToServer() {
+export function saveNotebookToServer(appMsg = true) {
   return (dispatch, getState) => {
     const state = getState();
     const notebookId = getNotebookID(state);
     const notebookInServer = Boolean(notebookId);
     if (notebookInServer) {
       // Update Exisiting Notebook
-      saveNotebookRequest(
+      return saveNotebookRequest(
         `/api/v1/notebooks/${notebookId}/revisions/`,
         getNotebookSaveRequestOptions(state),
-        dispatch
+        dispatch,
+        appMsg
       )
         .then(() => {
           const message = "Updated Notebook";
           updateAutosave(state, true);
-          dispatch(
-            updateAppMessages({
-              message,
-              messageType: "NOTEBOOK_SAVED"
-            })
-          );
+          if (appMsg) {
+            dispatch(
+              updateAppMessages({
+                message,
+                messageType: "NOTEBOOK_SAVED"
+              })
+            );
+          }
           dispatch({ type: "NOTEBOOK_SAVED" });
         })
-        .catch(() => {
-          // do nothing here.
+        .catch(err => {
+          throw Error(err);
         });
-    } else {
-      createNewNotebookOnServer()(dispatch, getState);
     }
+    return createNewNotebookOnServer()(dispatch, getState);
   };
 }
 
@@ -528,5 +549,18 @@ export function saveEnvironment(updateObj, update) {
     type: "SAVE_ENVIRONMENT",
     updateObj,
     update
+  };
+}
+
+export function setMostRecentSavedContent() {
+  return {
+    type: "SET_MOST_RECENT_SAVED_CONTENT"
+  };
+}
+
+export function setConnectionStatus(status) {
+  return {
+    type: "SET_CONNECTION_STATUS",
+    status
   };
 }