+++ src/state-schemas/state-schema.js
@@ -128,6 +128,14 @@ export const stateProperties = {
     type: "string",
     default: ""
   },
+  previouslySavedContent: {
+    type: "object",
+    default: { title: "", jsmd: "" },
+    properties: {
+      title: { type: "string" },
+      jsmd: { type: "string" }
+    }
+  },
   jsmdChunks: {
     type: "array",
     items: {
@@ -226,6 +234,10 @@ export const stateProperties = {
     properties: {
       user_can_save: { type: "boolean" },
       notebook_id: { type: "integer" },
+      connectionStatus: {
+        type: "string",
+        enum: ["CONNECTION_ACTIVE", "CONNECTION_LOST"]
+      },
       connectionMode: {
         type: "string",
         enum: ["SERVER", "STANDALONE"]
@@ -235,7 +247,8 @@ export const stateProperties = {
     default: {
       notebook_id: undefined,
       user_can_save: false,
-      connectionMode: "STANDALONE"
+      connectionMode: "STANDALONE",
+      connectionStatus: "CONNECTION_ACTIVE"
     }
   },
   panePositions: {